sync_music: Split metadata related code into own module
