+++ sync_music/transcode.py
@@ -17,7 +17,6 @@
 
 """Transcode action."""
 
-import base64
 import collections
 import logging
 
@@ -27,71 +26,37 @@ import mutagen
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
+ReplayGain = collections.namedtuple("ReplayGain", ["gain", "peak"])
 
-class Transcode:  # pylint: disable=too-many-instance-attributes
+
+class Transcode:
     """Transcodes audio files."""
 
-    def __init__(  # pylint: disable=too-many-arguments
+    name = "Transcoding"
+
+    def __init__(
         self,
         mode="auto",
         replaygain_preamp_gain=0.0,
-        transcode=True,
-        copy_tags=True,
-        albumartist_artist_hack=False,
-        albumartist_composer_hack=False,
-        artist_albumartist_hack=False,
-        discnumber_hack=False,
-        tracknumber_hack=False,
     ):
-        self.name = "Processing"
         self._format = audiotools.MP3Audio
         self._compression = "standard"
 
         logger.info("Transcoding settings:")
         logger.info(" - Audiotools %s", audiotools.VERSION)
-        logger.info(" - Mutagen %s", mutagen.version_string)
         self._mode = mode
-        self._transcode = transcode
-        if transcode and mode in [
-            "auto",
-            "transcode",
-            "replaygain",
-            "replaygain-album",
-        ]:
+
+        logger.info(
+            " - Converting to %s in quality %s",
+            self._format.NAME,
+            self._compression,
+        )
+        self._replaygain_preamp_gain = replaygain_preamp_gain
+        if mode.startswith("replaygain") and replaygain_preamp_gain != 0.0:
             logger.info(
-                " - Converting to %s in quality %s",
-                self._format.NAME,
-                self._compression,
+                " - Applying ReplayGain pre-amp gain %s", replaygain_preamp_gain
             )
-            self._replaygain_preamp_gain = replaygain_preamp_gain
-            if mode.startswith("replaygain") and replaygain_preamp_gain != 0.0:
-                logger.info(
-                    " - Applying ReplayGain pre-amp gain %s", replaygain_preamp_gain
-                )
-        else:
-            logger.info(" - Skipping transcoding")
 
-        self._copy_tags = copy_tags
-        if copy_tags:
-            logger.info(" - Copying tags")
-        else:
-            logger.info(" - Skipping copying tags")
-
-        self._albumartist_artist_hack = albumartist_artist_hack
-        if albumartist_artist_hack:
-            logger.info(" - Writing albumartist into artist field")
-        self._albumartist_composer_hack = albumartist_composer_hack
-        if albumartist_composer_hack:
-            logger.info(" - Writing albumartist into composer field")
-        self._artist_albumartist_hack = artist_albumartist_hack
-        if artist_albumartist_hack:
-            logger.info(" - Writing artist into albumartist field")
-        self._discnumber_hack = discnumber_hack
-        if discnumber_hack:
-            logger.info(" - Extending album field by disc number")
-        self._tracknumber_hack = tracknumber_hack
-        if tracknumber_hack:
-            logger.info(" - Remove track total from track number")
         logger.info("")
 
     @staticmethod
@@ -108,37 +73,6 @@ class Transcode:  # pylint: disable=too-many-instance-attributes
         return path.with_suffix(self.get_out_filetype())
 
     def execute(self, in_filepath, out_filepath):
-        """Executes action."""
-        if self._transcode:
-            self.transcode(in_filepath, out_filepath)
-
-        if self._copy_tags:
-            self.copy_tags(in_filepath, out_filepath)
-
-    def get_replaygain(self, in_filepath):
-        """Read ReplayGain info from tags."""
-        in_file = mutagen.File(in_filepath)
-        tag_prefix = "TXXX:" if isinstance(in_file, mutagen.mp3.MP3) else ""
-        rp_info = collections.namedtuple("ReplayGainInfo", ["gain", "peak"])
-        try:
-
-            def _get_value(tag):
-                value = in_file.tags[f"{tag_prefix}{tag}"][0]
-                return float(value.replace("dB", ""))
-
-            if self._mode == "replaygain-album":
-                return rp_info(
-                    _get_value("replaygain_album_gain"),
-                    _get_value("replaygain_album_peak"),
-                )
-
-            return rp_info(
-                _get_value("replaygain_track_gain"), _get_value("replaygain_track_peak")
-            )
-        except (TypeError, KeyError):
-            return None
-
-    def transcode(self, in_filepath, out_filepath):
         """Transcode audio file."""
         logger.info("Transcoding from %s to %s", in_filepath, out_filepath)
         try:
@@ -148,7 +82,7 @@ class Transcode:  # pylint: disable=too-many-instance-attributes
                 )
             else:
                 in_file = audiotools.open(str(in_filepath))
-                rp_info = self.get_replaygain(in_filepath)
+                rp_info = self._get_replaygain(in_filepath)
                 if rp_info:
                     pcmreader = audiotools.replaygain.ReplayGainReader(
                         in_file.to_pcm(),
@@ -166,200 +100,24 @@ class Transcode:  # pylint: disable=too-many-instance-attributes
         except (audiotools.EncodingError, audiotools.UnsupportedFile) as err:
             raise IOError(f"Failed to transcode file {in_filepath}: {err}") from err
 
-    def copy_tags(self, in_filepath, out_filepath):
-        """Copy tags."""
+    def _get_replaygain(self, in_filepath):
+        """Read ReplayGain info from tags."""
         in_file = mutagen.File(in_filepath)
-
-        # Tags are converted to ID3 format. If the output format is changed
-        # in the functions above, this function has to be adapted too.
+        tag_prefix = "TXXX:" if isinstance(in_file, mutagen.mp3.MP3) else ""
         try:
-            mp3_file = mutagen.mp3.MP3(out_filepath)
-        except mutagen.mp3.HeaderNotFoundError as err:
-            raise IOError("Output file is not in MP3 format") from err
 
-        if not mp3_file.tags:
-            mp3_file.tags = mutagen.id3.ID3()
+            def _get_value(tag):
+                value = in_file.tags[f"{tag_prefix}{tag}"][0]
+                return float(value.replace("dB", ""))
 
-        # Tags are processed depending on their input format.
-        if isinstance(in_file, mutagen.mp3.MP3):
-            self.copy_id3_to_id3(in_file.tags, mp3_file.tags)
-        elif isinstance(in_file, (mutagen.flac.FLAC, mutagen.oggvorbis.OggVorbis)):
-            self.copy_vorbis_to_id3(in_file.tags, mp3_file.tags)
-            self.copy_vorbis_picture_to_id3(in_file, mp3_file.tags)
-        else:
-            raise IOError("Input file tag conversion not implemented")
-
-        # Load the image from folder.jpg
-        self.copy_folder_image_to_id3(in_filepath, mp3_file.tags)
-
-        # Apply hacks
-        if self._albumartist_artist_hack:
-            self.apply_albumartist_artist_hack(mp3_file.tags)
-        if self._albumartist_composer_hack:
-            self.apply_albumartist_composer_hack(mp3_file.tags)
-        if self._artist_albumartist_hack:
-            self.apply_artist_albumartist_hack(mp3_file.tags)
-        if self._discnumber_hack:
-            self.apply_disknumber_hack(mp3_file.tags)
-        if self._tracknumber_hack:
-            self.apply_tracknumber_hack(mp3_file.tags)
-
-        # Remove ReplayGain tags if the volume has already been changed
-        if self._mode.startswith("replaygain"):
-            mp3_file.tags.delall("TXXX:replaygain_album_gain")
-            mp3_file.tags.delall("TXXX:replaygain_album_peak")
-            mp3_file.tags.delall("TXXX:replaygain_track_gain")
-            mp3_file.tags.delall("TXXX:replaygain_track_peak")
-
-        # Save as id3v1 and id3v2.3
-        mp3_file.tags.update_to_v23()
-        mp3_file.tags.save(out_filepath, v1=2, v2_version=3)
-
-    @classmethod
-    def copy_vorbis_to_id3(cls, src_tags, dest_tags):
-        """Copy tags in vorbis comments (ogg, flac) to ID3 format."""
-        tagtable = {
-            "album": mutagen.id3.TALB,
-            "artist": mutagen.id3.TPE1,
-            "albumartist": mutagen.id3.TPE2,
-            "title": mutagen.id3.TIT2,
-            "genre": mutagen.id3.TCON,
-            "date": mutagen.id3.TDRC,
-            "tracknumber": mutagen.id3.TRCK,
-            "discnumber": mutagen.id3.TPOS,
-            "MUSICBRAINZ_TRACKID": "http://musicbrainz.org",
-            "MUSICBRAINZ_ARTISTID": "MusicBrainz Artist Id",
-            "MUSICBRAINZ_ALBUMARTISTID": "MusicBrainz Album Artist Id",
-            "MUSICBRAINZ_RELEASEGROUPID": "MusicBrainz Release Group Id",
-            "MUSICBRAINZ_ALBUMID": "MusicBrainz Album Id",
-            "MUSICBRAINZ_RELEASETRACKID": "MusicBrainz Release Track Id",
-            "replaygain_album_gain": "replaygain_album_gain",
-            "replaygain_album_peak": "replaygain_album_peak",
-            "replaygain_track_gain": "replaygain_track_gain",
-            "replaygain_track_peak": "replaygain_track_peak",
-        }
-        for tag, id3tag in tagtable.items():
-            if tag in src_tags:
-                if tag == "tracknumber":
-                    track = src_tags["tracknumber"][0]
-                    if "tracktotal" in src_tags:
-                        track = f"{track}/{src_tags['tracktotal'][0]}"
-                    dest_tags.add(id3tag(encoding=3, text=track))
-                elif tag == "discnumber":
-                    disc = src_tags["discnumber"][0]
-                    if "disctotal" in src_tags:
-                        disc = f"{disc}/{src_tags['disctotal'][0]}"
-                    dest_tags.add(id3tag(encoding=3, text=disc))
-                elif tag == "MUSICBRAINZ_TRACKID":
-                    dest_tags.add(
-                        mutagen.id3.UFID(owner=id3tag, data=src_tags[tag][0].encode())
-                    )
-                elif isinstance(id3tag, str):  # TXXX tags
-                    dest_tags.add(
-                        mutagen.id3.TXXX(encoding=3, desc=id3tag, text=src_tags[tag])
-                    )
-                else:  # All other tags
-                    dest_tags.add(id3tag(encoding=3, text=src_tags[tag]))
-
-    @classmethod
-    def copy_vorbis_picture_to_id3(cls, in_file, dest_tags):
-        """Copy pictures from vorbis comments to ID3 format."""
-        pictures = []
-        try:  # Flac
-            pictures.extend(in_file.pictures)
-        except AttributeError:
-            pass
-
-        if "METADATA_BLOCK_PICTURE" in in_file.tags:  # OggVorbis
-            for data in in_file.tags["METADATA_BLOCK_PICTURE"]:
-                pictures.append(mutagen.flac.Picture(base64.b64decode(data)))
-        for picture in pictures:
-            dest_tags.add(
-                mutagen.id3.APIC(
-                    encoding=3,
-                    desc=picture.desc,
-                    data=picture.data,
-                    type=picture.type,
-                    mime=picture.mime,
+            if self._mode == "replaygain-album":
+                return ReplayGain(
+                    _get_value("replaygain_album_gain"),
+                    _get_value("replaygain_album_peak"),
                 )
+
+            return ReplayGain(
+                _get_value("replaygain_track_gain"), _get_value("replaygain_track_peak")
             )
-
-    @classmethod
-    def copy_id3_to_id3(cls, src_tags, dest_tags):
-        """Copy tags from ID3 to ID3."""
-        taglist = [
-            "TALB",
-            "TPE1",
-            "TPE2",
-            "TIT2",
-            "TCON",
-            "TDRC",
-            "TRCK",
-            "TPOS",
-            "APIC:",
-            "UFID:http://musicbrainz.org",
-            "TXXX:MusicBrainz Artist Id",
-            "TXXX:MusicBrainz Album Artist Id",
-            "TXXX:MusicBrainz Release Group Id",
-            "TXXX:MusicBrainz Album Id",
-            "TXXX:MusicBrainz Release Track Id",
-            "TXXX:replaygain_album_gain",
-            "TXXX:replaygain_album_peak",
-            "TXXX:replaygain_track_gain",
-            "TXXX:replaygain_track_peak",
-        ]
-        if src_tags is None:
-            return
-        for tag in taglist:
-            if tag in src_tags:
-                dest_tags.add(src_tags[tag])
-
-    @classmethod
-    def copy_folder_image_to_id3(cls, in_filename, dest_tags):
-        """Copy folder.jpg to ID3 tag."""
-        if "APIC:" not in dest_tags:
-            image = in_filename.parent / "folder.jpg"
-            if image.exists():
-                with image.open("rb") as image_file:
-                    img = image_file.read()
-                dest_tags.add(mutagen.id3.APIC(3, "image/jpg", 3, "", img))
-
-    @classmethod
-    def apply_albumartist_artist_hack(cls, tags):
-        """Copy the albumartist (TPE2) into the artist field (TPE1)."""
-        artist = tags["TPE2"].text if "TPE2" in tags else "Various Artists"
-        tags.add(mutagen.id3.TPE1(encoding=3, text=artist))
-
-    @classmethod
-    def apply_albumartist_composer_hack(cls, tags):
-        """Copy the albumartist (TPE2) into the composer field (TCOM)."""
-        if "TPE2" in tags:
-            tags.add(mutagen.id3.TCOM(encoding=3, text=tags["TPE2"].text))
-
-    @classmethod
-    def apply_artist_albumartist_hack(cls, tags):
-        """Copy the artist (TPE1) into the albumartist field (TPE2)."""
-        albumartist = tags["TPE1"].text if "TPE1" in tags else "Various Artists"
-        tags.add(mutagen.id3.TPE2(encoding=3, text=albumartist))
-
-    @classmethod
-    def apply_disknumber_hack(cls, tags):
-        """Extend album field by disc number."""
-        if "TALB" in tags and "TPOS" in tags and not tags["TPOS"] == "1":
-            tags.add(
-                mutagen.id3.TALB(
-                    encoding=tags["TALB"].encoding,
-                    text=tags["TALB"].text[0] + " - " + tags["TPOS"].text[0],
-                )
-            )
-
-    @classmethod
-    def apply_tracknumber_hack(cls, tags):
-        """Remove track total from track number."""
-        if "TRCK" in tags:
-            track_string = tags["TRCK"].text[0].split("/")[0]
-            try:
-                track_string = str(int(track_string))
-            except ValueError:
-                pass
-            tags.add(mutagen.id3.TRCK(encoding=0, text=track_string))
+        except (TypeError, KeyError):
+            return None
