+++ tests/test_transcode.py
@@ -18,152 +18,60 @@
 """Tests the transcoding implementation."""
 
 import pathlib
-import shutil
 
 import pytest
 
-from sync_music.sync_music import Transcode
+from sync_music.transcode import Transcode
+
+from tests import REFERENCE_FILES
 
 
 class TestTranscode:
     """Tests the transcoding implementation."""
 
-    in_filename_flac = "withtags.flac"
-    in_filename_flacall = "withalltags.flac"
-    in_filename_flacempty = "stripped.flac"
-    in_filename_ogg = "withtags.ogg"
-    in_filename_oggall = "withalltags.ogg"
-    in_filename_oggempty = "stripped.ogg"
-    in_filename_mp3 = "withtags.mp3"
-    in_filename_mp3all = "withalltags.mp3"
-    in_filename_mp3empty = "stripped.mp3"
-    in_filename_aiff = "withtags.aiff"
-    out_filename = "withtags.mp3"
-    img_filename = "folder.jpg"
-    input_path = pathlib.Path("tests/reference_data/audiofiles")
-    output_path = None
-
-    @pytest.fixture(autouse=True)
-    def init_output_path(self, tmp_path):
+    @staticmethod
+    @pytest.fixture
+    def out_path(tmp_path):
         """Setup temporary output directory."""
-        self.output_path = tmp_path
+        return tmp_path / "withtags.mp3"
 
-    def test_filename(self):
+    @staticmethod
+    def test_filename():
         """Tests retrieving the output file name."""
         transcode = Transcode()
-        out_filename = transcode.get_out_filename(pathlib.Path(self.in_filename_flac))
-        assert pathlib.Path(self.out_filename) == out_filename
+        out_filename = transcode.get_out_filename(REFERENCE_FILES.FLAC)
+        assert out_filename == REFERENCE_FILES.FLAC.with_suffix(".mp3")
 
-    def execute_transcode(
-        self, transcode, in_filename=in_filename_flac, out_filename=out_filename
-    ):
-        """Helper method to run transcoding tests."""
-        transcode.execute(
-            self.input_path / in_filename,
-            self.output_path / out_filename,
-        )
-
-    def test_transcode_default(self):
+    @staticmethod
+    def test_transcode_default(out_path):
         """Test transcoding with default options."""
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_mp3)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_mp3all)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_mp3empty)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_flac)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_flacall)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_flacempty)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_ogg)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_oggall)
-        self.execute_transcode(Transcode(), in_filename=self.in_filename_oggempty)
+        Transcode().execute(REFERENCE_FILES.MP3, out_path)
+        Transcode().execute(REFERENCE_FILES.FLAC, out_path)
+        Transcode().execute(REFERENCE_FILES.OGG, out_path)
 
-    def test_transcode_transcode(self):
-        """Tests transcoding with forced transcode."""
-        self.execute_transcode(
-            Transcode(mode="transcode"), in_filename=self.in_filename_mp3
-        )
-
-    def test_transcode_replaygain(self):
+    @staticmethod
+    def test_transcode_replaygain(out_path):
         """Tests transcoding with ReplayGain (track based)."""
-        self.execute_transcode(
-            Transcode(mode="replaygain"), in_filename=self.in_filename_mp3all
-        )
-        self.execute_transcode(
-            Transcode(mode="replaygain"), in_filename=self.in_filename_mp3
-        )
+        Transcode(mode="replaygain").execute(REFERENCE_FILES.MP3_ALL, out_path)
+        Transcode(mode="replaygain").execute(REFERENCE_FILES.MP3, out_path)
 
-    def test_transcode_replaygainalbum(self):
+    @staticmethod
+    def test_transcode_replaygainalbum(out_path):
         """Tests transcoding with ReplayGain (album based)."""
-        self.execute_transcode(
-            Transcode(mode="replaygain-album", replaygain_preamp_gain=10.0),
-            in_filename=self.in_filename_mp3,
+        Transcode(mode="replaygain-album", replaygain_preamp_gain=10.0).execute(
+            REFERENCE_FILES.MP3_ALL, out_path
         )
 
-    def test_transcode_folderimage(self):
-        """Tests transcoding without folder image."""
-        # Copy input file to a folder without folder.jpg (output folder)
-        in_filename = self.output_path / self.in_filename_flac
-        shutil.copy(self.input_path / self.in_filename_flac, in_filename)
-        self.execute_transcode(Transcode(), in_filename=in_filename)
-
-    def test_transcode_artist_hack(self):
-        """Tests transcoding with albumartist to artist hack enabled."""
-        self.execute_transcode(Transcode(albumartist_artist_hack=True))
-        self.execute_transcode(
-            Transcode(albumartist_artist_hack=True),
-            in_filename=self.in_filename_mp3empty,
-        )
-
-    def test_transcode_composer_hack(self):
-        """Tests transcoding with albumartist to composer hack enabled."""
-        self.execute_transcode(Transcode(albumartist_composer_hack=True))
-        self.execute_transcode(
-            Transcode(albumartist_composer_hack=True),
-            in_filename=self.in_filename_mp3empty,
-        )
-
-    def test_transcode_artist_albumartist_hack(self):
-        """Tests transcoding with albumartist to artist hack enabled."""
-        self.execute_transcode(Transcode(artist_albumartist_hack=True))
-        self.execute_transcode(
-            Transcode(artist_albumartist_hack=True),
-            in_filename=self.in_filename_mp3empty,
-        )
-
-    def test_transcode_discnumber_hack(self):
-        """Tests transcoding with disc number hack enabled."""
-        self.execute_transcode(Transcode(discnumber_hack=True))
-        self.execute_transcode(
-            Transcode(discnumber_hack=True), in_filename=self.in_filename_mp3all
-        )
-
-    def test_transcode_tracknumber_hack(self):
-        """Tests transcoding with track number hack enabled."""
-        self.execute_transcode(Transcode(tracknumber_hack=True))
-        self.execute_transcode(
-            Transcode(tracknumber_hack=True), in_filename=self.in_filename_mp3empty
-        )
-        self.execute_transcode(
-            Transcode(tracknumber_hack=True), in_filename="brokentag_tracknumber.mp3"
-        )
-
-    def test_transcodeerror_transcode(self):
+    @staticmethod
+    def test_transcode_transcodeerror():
         """Tests transcoding failure."""
         # IOError is raised on audiotools.EncodingError. The easiest way that
         # leads into this exception is writing to a non writable path.
         with pytest.raises(IOError):
-            self.execute_transcode(Transcode(), out_filename="/")
+            Transcode().execute(REFERENCE_FILES.FLAC, pathlib.Path("/"))
 
-    def test_transcodeerror_copytags(self):
-        """Tests copying tag failure."""
-        # Copy tags expects the output file to be an MP3 file.
-        # Don't transcode as this would rewrite the output file.
-        shutil.copy(
-            self.input_path / self.img_filename,
-            self.output_path / self.out_filename,
-        )
-        with pytest.raises(IOError):
-            self.execute_transcode(Transcode(transcode=False))
-
-    def test_transcodingerror_format(self):
+    @staticmethod
+    def test_transcode_formaterror(out_path):
         """Tests transcoding a non supported format."""
         with pytest.raises(IOError):
-            self.execute_transcode(Transcode(), in_filename=self.in_filename_aiff)
+            Transcode().execute(REFERENCE_FILES.FOLDER_IMAGE, out_path)
